requirements.txt
git+https://github.com/nousresearch/hermes-agent.git
openai>=0.27.0
openrouter>=0.0.4
apify-client>=1.0.0
pdfplumber>=0.7.6
pytesseract>=0.3.9
pypdf>=3.11.0
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
